# 상품 취소, 예약 시간 등록 등

+++ src/main/java/com/knocksea/see/edu/entity/Edu.java
@@ -23,7 +23,7 @@ public class Edu {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(length =10)
-    private int eduId;
+    private Long eduId;
 
     @Column(updatable = false)
     @CreationTimestamp
